requirements.txt
deepeval
mcp
openai
anthropic
google-generativeai
google-genai
